@@ -1,7 +1,7 @@
 'use strict';
 
-import React from 'react-native';
+import ReactNative from 'react-native';
 
 import App from './src/App';
 
-React.AppRegistry.registerComponent('CineHorarios', function() { return App });
+ReactNative.AppRegistry.registerComponent('CineHorarios', function() { return App });
