@@ -1,6 +1,7 @@
 'use strict';
 
-import React, {
+import React from 'react';
+import {
   ActivityIndicatorIOS,
   Platform,
   ProgressBarAndroid,
