@@ -1,6 +1,6 @@
 'use strict';
 
-import React from 'react-native'
+import React from 'react';
 
 export default function renderNormalScene(route, navigator) {
   const { title, Component, extraData } = route;
