@@ -1,10 +1,8 @@
 'use strict';
 
-import React, {
-  PropTypes
-} from 'react-native'
+import React, { PropTypes } from 'react'
 
-import {MyGiftedListView} from '../ReusableComponents'
+import { MyGiftedListView } from '../ReusableComponents'
 import MenuCell from './componentMenuCell'
 
 export default class ComponentMenu extends React.Component {
