@@ -1,6 +1,7 @@
 'use strict';
 
-import React, { Image, StyleSheet } from 'react-native'
+import React from 'react';
+import { Image, StyleSheet } from 'react-native'
 
 export default class RightAccesoryView extends React.Component {
 
