@@ -1,10 +1,10 @@
 'use strict';
 
-export { default as MyGiftedListView } from './MyGiftedListView'
-export { default as EmptyView } from './EmptyView'
-export { default as MovieCell } from './MovieCell'
-export { default as RightAccessoryView } from './RightAccessoryView'
-export { default as SeparatorView } from './SeparatorView'
-export { default as ResponsiveImage } from './ResponsiveImage'
-export { default as SimpleCell } from './SimpleCell'
-export { default as MyListViewCell } from './MyListViewCell'
+export { default as MyGiftedListView } from './MyGiftedListView';
+export { default as EmptyView } from './EmptyView';
+export { default as MovieCell } from './MovieCell';
+export { default as RightAccessoryView } from './RightAccessoryView';
+export { default as SeparatorView } from './SeparatorView';
+export { default as ResponsiveImage } from './ResponsiveImage';
+export { default as SimpleCell } from './SimpleCell';
+export { default as MyListViewCell } from './MyListViewCell';
