@@ -1,7 +1,7 @@
 'use strict';
 
-import React, {
-  PropTypes,
+import React, { PropTypes } from 'react';
+import {
   View,
   StyleSheet,
   Text,
