@@ -1,25 +1,23 @@
 'use strict';
 
-import React, {
+import React, { PropTypes } from 'react';
+import {
   StyleSheet,
   TouchableHighlight,
   View,
   Image,
-  Text,
-  PropTypes
+  Text
 } from 'react-native'
 
-import { colors } from '../../Data'
-import { Api, ImageHelper } from '../../Utils'
+import { colors } from '../../Data';
 
 export default class ComponentVideoCell extends React.Component {
 
   static propTypes = {
-  	title: PropTypes.string,
-  	subtitle: PropTypes.string,
-  	image1: PropTypes.string,
-  	image2: PropTypes.string,
-  	rowID: PropTypes.string,
+    videoName: PropTypes.string,
+    videoPortraitImageUri: PropTypes.string,
+    showName: PropTypes.string,
+    showImageUri: PropTypes.string,
     onPress: PropTypes.func,
     onPressShow: PropTypes.func
   };
@@ -30,23 +28,31 @@ export default class ComponentVideoCell extends React.Component {
 	    <View style={styles.rowContainer}>
 
 	      <TouchableHighlight
+          style={[styles.showContainer, styles.shadow]}
 	        underlayColor={colors.concrete}
 	        onPress={() => this.props.onPress() }>
 	        <Image
-		        source={{uri: Api.getFullURL(ImageHelper.getThumbImage(this.props.image1))}}
 		        style={styles.showImage}
+            source={{uri: this.props.showImageUri}}
 		       />
 	      </TouchableHighlight>
 
 
 	      <TouchableHighlight
+          style={[styles.videoContainer, styles.shadow]}
 	        underlayColor={colors.concrete}
 	        onPress={() => this.props.onPressShow() }>
 	        <Image
 	          resizeMode='cover'
 	          style={styles.portraitImage}
-	          source={{uri: Api.getFullURL(ImageHelper.getThumbImage(this.props.image2))}}
-	         />
+            source={{uri: this.props.videoPortraitImageUri}}
+	        >
+            <View style={styles.videoBackground}>
+              <Image
+                source={require('../../../assets/VIdeoPlay.png')}
+              />
+            </View>
+          </Image>
 	      </TouchableHighlight>
 
 	    </View>
@@ -62,16 +68,32 @@ let styles = StyleSheet.create({
     padding: 10,
     backgroundColor: '#FFF'
   },
+  showContainer: {
+  },
+  videoContainer: {
+    flex: 1,
+    marginLeft: 10,
+  },
+  shadow: {
+    shadowColor: 'black',
+    shadowRadius: 3,
+    shadowOpacity: 1,
+    shadowOffset: {
+      width: 0, height: 0
+    },
+    backgroundColor: 'gray'
+  },
   showImage: {
-    flex: 2,
     width: 80,
     height: 120,
-    marginRight: 10
   },
   portraitImage: {
     flex: 1,
-    width: 200,
-    height: 120,
     backgroundColor: 'black'
+  },
+  videoBackground: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.2)',
+    alignItems: 'center'
   }
 });
